Share implementation of ImplementAbstractClass.
