+++ src/Features/Core/Portable/FeaturesResources.resx
@@ -1061,4 +1061,7 @@ This version used in: {2}</value>
   <data name="paren_Unknown_paren" xml:space="preserve">
     <value>(Unknown)</value>
   </data>
+  <data name="Implement_Abstract_Class" xml:space="preserve">
+    <value>Implement Abstract Class</value>
+  </data>
 </root>
